requirements.txt
flask
flask-cors
gunicorn
requests
google-api-python-client
google-auth-oauthlib
edge-tts
